apps/sandbox/pyproject.toml
[tool.poetry]
name = "k8s-sandbox"
version = "0.1.0"
description = ""
authors = ["lukejagg <lukejagg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
uvicorn = "^0.31.1"
python-socketio = {extras = ["client"], version = "^5.11.4"}
fastapi-socketio = "^0.0.10"


[tool.poetry.group.dev.dependencies]
kubernetes = "^31.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
